# perf(expr): optimize casting to varchar