--- a/src/common/src/array/utf8_array.rs
+++ b/src/common/src/array/utf8_array.rs
@@ -159,12 +159,6 @@ pub struct StringWriter<'a> {
 }
 
 impl<'a> StringWriter<'a> {
-    /// `write_ref` will consume `StringWriter` and pass the ownership of `builder` to `BytesGuard`.
-    #[inline]
-    pub fn write_ref(self, value: &str) -> WrittenGuard {
-        self.bytes.write_ref(value.as_bytes())
-    }
-
     /// `begin` will create a `PartialStringWriter`, which allow multiple appendings to create a new
     /// record.
     pub fn begin(self) -> PartialStringWriter<'a> {
@@ -181,14 +175,6 @@ pub struct PartialStringWriter<'a> {
 }
 
 impl<'a> PartialStringWriter<'a> {
-    /// `write_ref` will append partial dirty data to `builder`.
-    /// `PartialStringWriter::write_ref` is different from `StringWriter::write_ref`
-    /// in that it allows us to call it multiple times.
-    #[inline]
-    pub fn write_ref(&mut self, value: &str) {
-        self.bytes.write_ref(value.as_bytes());
-    }
-
     /// `finish` will be called while the entire record is written.
     /// Exactly one new record was appended and the `builder` can be safely used.
     pub fn finish(self) -> WrittenGuard {
@@ -198,7 +184,7 @@ impl<'a> PartialStringWriter<'a> {
 
 impl Write for PartialStringWriter<'_> {
     fn write_str(&mut self, s: &str) -> std::fmt::Result {
-        self.write_ref(s);
+        self.bytes.write_ref(s.as_bytes());
         Ok(())
     }
 }
@@ -232,7 +218,7 @@ mod tests {
             let writer = builder.writer();
             let mut partial_writer = writer.begin();
             for _ in 0..2 {
-                partial_writer.write_ref("ran");
+                partial_writer.write_str("ran").unwrap();
             }
             partial_writer.finish()
         };
@@ -249,8 +235,8 @@ mod tests {
         let _guard: WrittenGuard = {
             let writer = builder.writer();
             let mut partial_writer = writer.begin();
-            partial_writer.write_ref("Dia");
-            partial_writer.write_ref("na");
+            partial_writer.write_str("Dia").unwrap();
+            partial_writer.write_str("na").unwrap();
             partial_writer.finish()
         };
 
@@ -258,8 +244,8 @@ mod tests {
         let _maybe_guard: Option<WrittenGuard> = {
             let writer = builder.writer();
             let mut partial_writer = writer.begin();
-            partial_writer.write_ref("Ca");
-            partial_writer.write_ref("rol");
+            partial_writer.write_str("Ca").unwrap();
+            partial_writer.write_str("rol").unwrap();
 
             // We don't finish here.
             None
@@ -269,8 +255,8 @@ mod tests {
         let _guard: WrittenGuard = {
             let writer = builder.writer();
             let mut partial_writer = writer.begin();
-            partial_writer.write_ref("Ki");
-            partial_writer.write_ref("ra");
+            partial_writer.write_str("Ki").unwrap();
+            partial_writer.write_str("ra").unwrap();
             partial_writer.finish()
         };