--- a/src/common/src/types/chrono_wrapper.rs
+++ b/src/common/src/types/chrono_wrapper.rs
@@ -72,39 +72,39 @@ impl Default for NaiveDateTimeWrapper {
 }
 
 impl ToText for NaiveDateWrapper {
-    fn fmt(&self, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
+    fn write(&self, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
         write!(f, "{}", self.0)
     }
 
-    fn to_text_with_type(&self, ty: &DataType) -> String {
+    fn write_with_type(&self, ty: &DataType, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
         match ty {
-            super::DataType::Date => self.to_text(),
+            super::DataType::Date => self.write(f),
             _ => unreachable!(),
         }
     }
 }
 
 impl ToText for NaiveTimeWrapper {
-    fn fmt(&self, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
+    fn write(&self, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
         write!(f, "{}", self.0)
     }
 
-    fn to_text_with_type(&self, ty: &DataType) -> String {
+    fn write_with_type(&self, ty: &DataType, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
         match ty {
-            super::DataType::Time => self.to_text(),
+            super::DataType::Time => self.write(f),
             _ => unreachable!(),
         }
     }
 }
 
 impl ToText for NaiveDateTimeWrapper {
-    fn fmt(&self, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
+    fn write(&self, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
         write!(f, "{}", self.0)
     }
 
-    fn to_text_with_type(&self, ty: &DataType) -> String {
+    fn write_with_type(&self, ty: &DataType, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
         match ty {
-            super::DataType::Timestamp => self.to_text(),
+            super::DataType::Timestamp => self.write(f),
             _ => unreachable!(),
         }
     }