--- a/src/common/src/types/decimal.rs
+++ b/src/common/src/types/decimal.rs
@@ -42,8 +42,8 @@ pub enum Decimal {
 }
 
 impl ToText for Decimal {
-    fn to_text(&self) -> String {
-        self.to_string()
+    fn fmt(&self, f: &mut dyn std::fmt::Write) -> std::fmt::Result {
+        write!(f, "{self}")
     }
 
     fn to_text_with_type(&self, ty: &DataType) -> String {