--- a/src/expr/src/vector_op/ltrim.rs
+++ b/src/expr/src/vector_op/ltrim.rs
@@ -21,7 +21,8 @@ use crate::Result;
 /// Since we would like to simplify the implementation, currently we omit this case.
 #[inline(always)]
 pub fn ltrim(s: &str, writer: &mut dyn Write) -> Result<()> {
-    Ok(writer.write_str(s.trim_start()).unwrap())
+    writer.write_str(s.trim_start()).unwrap();
+    Ok(())
 }
 
 #[cfg(test)]