… src/expr/src/expr/expr_concat_ws.rs
… src/expr/src/expr/expr_unary.rs
… src/expr/src/expr/template.rs
… src/expr/src/vector_op/cast.rs
… src/expr/src/vector_op/concat_op.rs
… src/expr/src/vector_op/lower.rs
… src/expr/src/vector_op/md5.rs
… src/expr/src/vector_op/overlay.rs
… src/expr/src/vector_op/repeat.rs
… src/expr/src/vector_op/replace.rs
… src/expr/src/vector_op/rtrim.rs
… src/expr/src/vector_op/substr.rs
… src/expr/src/vector_op/to_char.rs
… src/expr/src/vector_op/translate.rs
… src/expr/src/vector_op/trim.rs
… src/expr/src/vector_op/trim_characters.rs
… src/expr/src/vector_op/upper.rs